refactor: split build and task pod monitor controllers
