+++ internal/controllers/v1beta2/buildpodmonitor_metrics.go
@@ -9,7 +9,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
-func (r *LagoonMonitorReconciler) calculateBuildMetrics(ctx context.Context) error {
+func (r *BuildMonitorReconciler) calculateBuildMetrics(ctx context.Context) error {
 	listOption := (&client.ListOptions{}).ApplyOptions([]client.ListOption{
 		client.MatchingLabels(map[string]string{
 			"lagoon.sh/jobType":    "build",
@@ -27,26 +27,5 @@ func (r *LagoonMonitorReconciler) calculateBuildMetrics(ctx context.Context) err
 		}
 	}
 	metrics.BuildsRunningGauge.Set(runningBuilds)
-	return nil
-}
-
-func (r *LagoonMonitorReconciler) calculateTaskMetrics(ctx context.Context) error {
-	listOption := (&client.ListOptions{}).ApplyOptions([]client.ListOption{
-		client.MatchingLabels(map[string]string{
-			"lagoon.sh/jobType":    "task",
-			"lagoon.sh/controller": r.ControllerNamespace,
-		}),
-	})
-	taskPods := &corev1.PodList{}
-	if err := r.List(ctx, taskPods, listOption); err != nil {
-		return fmt.Errorf("unable to list tasks in the cluster, there may be none or something went wrong: %v", err)
-	}
-	runningTasks := float64(0)
-	for _, taskPod := range taskPods.Items {
-		if taskPod.Status.Phase == corev1.PodRunning {
-			runningTasks = runningTasks + 1
-		}
-	}
-	metrics.TasksRunningGauge.Set(runningTasks)
 	return nil
 }
