+++ internal/controllers/v1beta2/helpers.go
@@ -1,0 +1,44 @@
+package v1beta2
+
+import (
+	"bytes"
+	"context"
+	"fmt"
+	"io"
+
+	corev1 "k8s.io/api/core/v1"
+	"k8s.io/client-go/kubernetes"
+	ctrl "sigs.k8s.io/controller-runtime"
+	"sigs.k8s.io/controller-runtime/pkg/client/config"
+)
+
+// slice of the different failure states of pods that we care about
+// if we observe these on a pending pod, fail the build and get the logs
+var failureStates = []string{
+	"CrashLoopBackOff",
+	"ImagePullBackOff",
+}
+
+// getContainerLogs grabs the logs from a given container
+func getContainerLogs(ctx context.Context, containerName string, request ctrl.Request) ([]byte, error) {
+	restCfg, err := config.GetConfig()
+	if err != nil {
+		return nil, fmt.Errorf("unable to get config: %v", err)
+	}
+	clientset, err := kubernetes.NewForConfig(restCfg)
+	if err != nil {
+		return nil, fmt.Errorf("unable to create client: %v", err)
+	}
+	req := clientset.CoreV1().Pods(request.Namespace).GetLogs(request.Name, &corev1.PodLogOptions{Container: containerName})
+	podLogs, err := req.Stream(ctx)
+	if err != nil {
+		return nil, fmt.Errorf("error in opening stream: %v", err)
+	}
+	defer podLogs.Close()
+	buf := new(bytes.Buffer)
+	_, err = io.Copy(buf, podLogs)
+	if err != nil {
+		return nil, fmt.Errorf("error in copy information from podLogs to buffer: %v", err)
+	}
+	return buf.Bytes(), nil
+}
