+++ internal/controllers/v1beta2/task_helpers.go
@@ -11,7 +11,6 @@ import (
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/uselagoon/machinery/api/schema"
 	lagooncrd "github.com/uselagoon/remote-controller/api/lagoon/v1beta2"
-	"github.com/uselagoon/remote-controller/internal/helpers"
 	"github.com/uselagoon/remote-controller/internal/metrics"
 
 	rbacv1 "k8s.io/api/rbac/v1"
@@ -46,23 +45,6 @@ func (r *LagoonTaskReconciler) createActiveStandbyRole(ctx context.Context, sour
 		if err := r.Create(ctx, activeStandbyRoleBinding); err != nil {
 			return fmt.Errorf("there was an error creating the lagoon-deployer-activestandby role binding. Error was: %v", err)
 		}
-	}
-	return nil
-}
-
-// deleteActiveStandbyRole
-func (r *LagoonMonitorReconciler) deleteActiveStandbyRole(ctx context.Context, destinationNamespace string) error {
-	activeStandbyRoleBinding := &rbacv1.RoleBinding{}
-	err := r.Get(ctx, types.NamespacedName{
-		Namespace: destinationNamespace,
-		Name:      "lagoon-deployer-activestandby",
-	}, activeStandbyRoleBinding)
-	if err != nil {
-		helpers.IgnoreNotFound(err)
-	}
-	err = r.Delete(ctx, activeStandbyRoleBinding)
-	if err != nil {
-		return fmt.Errorf("unable to delete lagoon-deployer-activestandby role binding")
 	}
 	return nil
 }
